Add logging to telegram, fix bags

+++ handlers/my_cars.py
@@ -1,4 +1,6 @@
 """Работа с разделом МОИ АВТОМОБИЛИ"""
+import logging
+
 from aiogram import types, Dispatcher
 from aiogram.dispatcher import FSMContext
 
@@ -7,6 +9,8 @@ from keyboards.menu_bot import back_key, menu
 from misc.states import FSMActions
 from utils import db
 from utils.exceptions import NotCorrectCarName
+
+logger = logging.getLogger('telegram_logger')
 
 
 async def menu_cars_cmd(call: types.CallbackQuery):
@@ -27,8 +31,12 @@ async def new_car_handler(m: types.Message, state: FSMContext):
     async with state.proxy() as context_data:
         prev_m = context_data['prev_m']
     try:
-        db.add_new_car(str(m.from_user.id), m.text)
-        await prev_m.edit_text('Добавлен новый автомобиль!\n\n' + m.text)
+        car_name = m.text
+        if len(car_name) > 60:
+            raise NotCorrectCarName('Слишком длинное название!\nПопробуй еще раз')
+        db.add_new_car(str(m.from_user.id), car_name)
+        await prev_m.edit_text('Добавлен новый автомобиль!\n\n' + car_name)
+        logger.info(f'{m.from_user.first_name} добавил новый авто: {car_name}')
         await prev_m.edit_reply_markup(reply_markup=menu)
         await m.delete()
         await state.finish()
@@ -45,8 +53,10 @@ async def deleting_car(call: types.CallbackQuery):
 
 
 async def choice_car_to_delete(call: types.CallbackQuery, callback_data: dict):
-    db.delete_car(str(call.from_user.id), callback_data.get('car'))
+    car = callback_data.get('car')
+    db.delete_car(str(call.from_user.id), car)
     await call.message.edit_text(f'Автомобиль {callback_data.get("car")} удален!')
+    logger.info(f'{call.from_user.first_name} удалил автомобиль: {car}')
     await call.message.edit_reply_markup(reply_markup=menu)
